feat(component): refactor Select
