+++ src/pages/ExamplePage/ExamplePage.tsx
@@ -2,7 +2,7 @@ import { Container, Divider, Stack, Typography } from '@mui/material';
 import MuiComponents from './components/MuiComponents';
 import ModalComponents from './components/ModalComponents';
 import ReactQueryComponent from './components/ReactQueryComponent';
-import SelectorComponents from './components/SelectorComponents';
+import SelectComponents from './components/SelectComponents';
 import SwiperComponent from './components/SwiperComponent';
 import Footer from '@src/common/Footer';
 import HotelTypography from './components/HotelTypography';
@@ -29,8 +29,8 @@ const ExamplePage = () => {
           <Divider textAlign="center">Modal</Divider>
           <ModalComponents />
 
-          <Divider textAlign="center">Selector</Divider>
-          <SelectorComponents />
+          <Divider textAlign="center">Select</Divider>
+          <SelectComponents />
           
         </Stack>
       </Container>
