--- src/ui-components/Selector/index.ts
@@ -1,1 +1,0 @@
-export { default } from './Selector';
